Get version info into Docker image
Version data comes from `git describe`, which requires `git` to be
installed as well as the git repo! Good thing about Go is we can have a
throw-away build stage, so the image will never have `.git` in it.

Even so, long-term the dockerfile really should just do a git clone so
there's no chance of accidentally including sensitive files.